GY150331
The original version is not fully compliant with the GTF specification. Separately, it fails to correctly parse GTF data directly downloaded from Ensembl, exiting with an error.

The script also retains features from, although not support for Python2 versions back to 2.3 which dates from 2003. This is at the expense of supporting Python3 and even later Python2 idioms. It should be noted that the system default Python is now 3.x for some Linux distributions, so the hashbang '#!/usr/bin/env python'  should either be changed to '#!/usr/bin/env python2' or the file should run without changes under Python3.

With the latter point in mind, the script has been updated to support both Python 2.7 & 3+. It now has a more useful help message interface, as well as support for parsing data from pipes (allowing GTF files merged with cat to be passed). Various efficiency improvements have been made and the code has been re-factored to bring it in line with PEP8. Overall, the naming systems and design of the code have not been greatly altered.

NB: A number of assertions have been removed, as they were largely arbitrary.  It would be simple to re-insert these if preferred.